Prototype implementation of keepalives
This is a possible implementation of keepalives for Issue #854.

Add the command-line option "/send-keepalives", which has an optional
argument:

  always:   Always send keepalives (the default if specified without an
            argument).
  never:    Don't send keepalives (the default if option not given).
  ifactive: Send keepalives only while the client desktop session is
            itself active.

Add functions input_send_fastpath_keepalive_event() and
input_send_keepalive_event(), which are added to the SendKeepaliveEvent
field in rdpInput.  These functions send a keepalive, which is currently
simply a mouse movement to the last (x, y) pointer coordinates
previously sent.  (Mouse movement coordinates are absolute, not
relative, so these coordinates need to be saved by each function that
sends mouse events.)

The keepalive is sent by these functions only if they are called after
KEEPALIVE_INTERVAL_MILLISECS milliseconds (by default 2 minutes) of the
previous keepalive message (a pointer to the time is passed in by the
caller).

The SendKeepaliveEvent member function is called from the
xf_client_thread() event loop.

Limitations:

1. This is only implemented for the X11 client.

2. Re-sending the current coordinates should work for a Windows server
but doesn't work if the server is xrdp, because it ignores the input
event as it hasn't moved.  I don't yet know if that is an xrdp or an X11
issue.  It can be worked around by changing input_send_keepalive_event()
and input_send_fastpath_keepalive_event() to send two events at (x+1, y)
and then (x, y); but that implies an actual mouse movement which is more
invasive so might cause problems for some applications such as games.

3. This should probably disable sending the suppress output event
corresponding to UnmapNotify (e.g. minimise).

4. I haven't yet implemented the "ifactive" option.  This could be
implemented for X11 using the X11 Screen Saver extension (see xprintidle
for an example of how this is used).

5. If real input events are sent then the event loop can reset the
keepalive timer; I haven't implemented that.